+++ packages/react/src/lib/providers/f0/f0-provider.tsx
@@ -17,6 +17,7 @@ import { LinkContextValue, LinkProvider } from "../../linkHandler"
 import { PrivacyModeProvider } from "../../privacyMode"
 import { cn } from "../../utils"
 import { XRayProvider } from "../../xray"
+import { DialogsAlikeLayoutProvider } from "../dialogs-alike/DialogsAlikeLayoutProvider"
 import { DataCollectionStorageProvider } from "../datacollection/DataCollectionStorageProvider"
 import { DataCollectionStorageHandler } from "../datacollection/types"
 import { I18nProvider, I18nProviderProps } from "../i18n"
@@ -144,11 +145,13 @@ export const F0Provider: React.FC<{
                       <DataCollectionStorageProvider
                         handler={dataCollectionStorageHandler}
                       >
-                        <FormComponentContext.Provider value={formComponent}>
-                          <FormCardValueFormatterProvider>
-                            {children}
-                          </FormCardValueFormatterProvider>
-                        </FormComponentContext.Provider>
+                        <DialogsAlikeLayoutProvider>
+                          <FormComponentContext.Provider value={formComponent}>
+                            <FormCardValueFormatterProvider>
+                              {children}
+                            </FormCardValueFormatterProvider>
+                          </FormComponentContext.Provider>
+                        </DialogsAlikeLayoutProvider>
                       </DataCollectionStorageProvider>
                     </ImageProvider>
                   </PrivacyModeProvider>